backend/requirements.txt
fastapi
uvicorn
python-dotenv
mistralai
python-multipart
reportlab
pymupdf
httpx
motor